+++ libs/base_core/src/media/color/mod.rs
@@ -1,0 +1,25 @@
+// devela_base_core::media::color
+//
+//! Chromatic functionality.
+//
+// safety
+#![cfg_attr(base_safe_color, forbid(unsafe_code))]
+
+mod helpers; // (Norm)
+
+mod gamma; // Gamma
+mod luminance; // Luma, Luminance
+mod rgb; // Rgb[a][8|16|F32|F64], Rgb[a]Lin[F32|F64]
+
+crate::structural_mods! { // _mods, _crate_internals
+    _mods {
+        pub use super::{
+            gamma::*,
+            luminance::*,
+            rgb::_all::*,
+        };
+    }
+    _crate_internals {
+        pub(crate) use super::helpers::*;
+    }
+}