+++ libs/base_core/src/media/color/rgb/impls/mod.rs
@@ -1,0 +1,12 @@
+// devela::media::color::rgb::impls
+//
+//!
+//
+
+// methods:
+mod u8;
+mod u8alpha;
+mod u16;
+mod u16alpha;
+crate::items! { mod f32; mod f32alpha; }
+crate::items! { mod f64; mod f64alpha; }