+++ libs/base_core/src/media/mod.rs
@@ -1,7 +1,7 @@
 // devela_base_core::media
 //
 #![doc = crate::_DOC_MEDIA!()]
-#![doc = crate::_doc!(modules: crate; media: audio, image)] // color, draw, font, video
+#![doc = crate::_doc!(modules: crate; media: audio, color, image)] // draw, font, video
 #![doc = crate::_doc!(newline)]
 //
 // safety
@@ -13,9 +13,9 @@
 #[cfg_attr(nightly_doc, doc(cfg(feature = "audio")))]
 pub mod audio;
 
-// #[cfg(feature = "color")]
-// #[cfg_attr(nightly_doc, doc(cfg(feature = "color")))]
-// pub mod color;
+#[cfg(feature = "color")]
+#[cfg_attr(nightly_doc, doc(cfg(feature = "color")))]
+pub mod color;
 
 // #[cfg(feature = "draw")]
 // #[cfg_attr(nightly_doc, doc(cfg(feature = "draw")))]
@@ -25,12 +25,13 @@ pub mod audio;
 #[cfg_attr(nightly_doc, doc(cfg(feature = "image")))]
 pub mod image;
 
-crate::structural_mods! { // _pub_mods
+crate::structural_mods! { // _pub_mods, _crate_internals
     _pub_mods {
         #[cfg(feature = "audio")]
         pub use super::audio::_all::*;
-        // #[cfg(feature = "color")]
-        // pub use super::color::_all::*;
+
+        #[cfg(feature = "color")]
+        pub use super::color::_all::*;
 
         // #[cfg(feature = "draw")]
         // pub use super::draw::_all::*;
@@ -38,5 +39,9 @@ crate::structural_mods! { // _pub_mods
         #[cfg(feature = "image")]
         #[cfg_attr(nightly_doc, doc(cfg(feature = "image")))]
         pub use super::image::_all::*;
+    }
+    _crate_internals {
+        #[cfg(feature = "color")]
+        pub use super::color::_crate_internals::*;
     }
 }