+++ src/media/color/color.rs
@@ -2,6 +2,15 @@
 //
 //! Defines the [`Color`] trait.
 //
+// TOC
+// - trait Color
+// - macro impl_color!
+// - impl Color
+
+use crate::{
+    Rgb8, Rgb16, RgbF32, RgbF64, RgbLinF32, RgbLinF64, Rgba8, Rgba16, RgbaF32, RgbaF64, RgbaLinF32,
+    RgbaLinF64, RgbaLinPreF32, RgbaLinPreF64, RgbaPre8, RgbaPre16, RgbaPreF32, RgbaPreF64,
+};
 
 use crate::{NotEnoughSpace, NumConst};
 #[cfg(feature = "alloc")]
@@ -104,4 +113,153 @@ pub trait Color {
         let _ = self.color_components_write(&mut buffer); // CHECK
         buffer
     }
+}
+
+/// Args
+/// - `$C` the number of components in the inner c array.
+/// - `$C` the numberof
+macro_rules! impl_color {
+    (
+    // rgb colors (3 components)
+    // - $Name   : the name of the rgb color type
+    // - $C      : the type of the inner component
+    // - $BITS   : the number of bits of each inner component
+    // - $INT    : a boolean indicating whether the components are integers
+    // - $LINEAR : a boolean indicating whether it's linear
+    rgb: $Name:ty, $C:ty, $BITS:literal, $INT:literal, $LINEAR:literal) => {
+        #[rustfmt::skip]
+        impl $crate::Color for $Name {
+            type Component = $C;
+            const COLOR_BITS: usize = $BITS;
+            const COLOR_COUNT: usize = 3;
+            const COLOR_HAS_ALPHA: bool = false;
+            const COLOR_IS_LINEAR: bool = $LINEAR;
+            const COLOR_IS_INT: bool = $INT;
+            const COLOR_IS_PREMUL: bool = false;
+
+            fn color_red(&self) -> Self::Component { self.c[0] }
+            fn color_green(&self) -> Self::Component { self.c[1] }
+            fn color_blue(&self) -> Self::Component { self.c[2] }
+            /// Since the color has no alpha, the maximum normalized value is returned.
+            fn color_alpha(&self) -> Self::Component {
+                // NOTE: there's a NUM_MAX_NORM for all supported primitive types
+                <Self::Component as $crate::NumConst>::NUM_MAX_NORM.unwrap()
+            }
+
+            fn color_components_write(&self, b: &mut [$C]) -> Result<(), $crate::NotEnoughSpace> {
+                let c = self.c;
+                let needed = Self::COLOR_COUNT;
+                if b.len() < needed {
+                    Err($crate::NotEnoughSpace(Some(Self::COLOR_COUNT)))
+                } else {
+                    b[0] = c[0];
+                    b[1] = c[1];
+                    b[2] = c[2];
+                    Ok(())
+                }
+            }
+        }
+    };
+    (
+    // rgba colors (4 components)
+    // - $Name   : the name of the rgba color type
+    // - $C      : the type of the inner component
+    // - $BITS   : the number of bits of each inner component
+    // - $INT    : a boolean indicating whether the components are integers
+    // - $LINEAR : a boolean indicating whether it's linear
+    // - $PREMUL : a boolean indicating whether the alpha is premultiplied
+    rgba: $Name:ty, $C:ty, $BITS:literal, $INT:literal, $LINEAR:literal, $PREMUL:literal) => {
+        #[rustfmt::skip]
+        impl $crate::Color for $Name {
+            type Component = $C;
+            const COLOR_BITS: usize = $BITS;
+            const COLOR_COUNT: usize = 4;
+            const COLOR_HAS_ALPHA: bool = true;
+            const COLOR_IS_LINEAR: bool = $LINEAR;
+            const COLOR_IS_INT: bool = $INT;
+            const COLOR_IS_PREMUL: bool = $PREMUL;
+
+            fn color_red(&self) -> Self::Component { self.c[0] }
+            fn color_green(&self) -> Self::Component { self.c[1] }
+            fn color_blue(&self) -> Self::Component { self.c[2] }
+            fn color_alpha(&self) -> Self::Component { self.c[3] }
+
+            fn color_components_write(&self, b: &mut [$C]) -> Result<(), $crate::NotEnoughSpace> {
+                let c = self.c;
+                let needed = Self::COLOR_COUNT;
+                if b.len() < needed {
+                    Err($crate::NotEnoughSpace(Some(Self::COLOR_COUNT)))
+                } else {
+                    b[0] = c[0];
+                    b[1] = c[1];
+                    b[2] = c[2];
+                    b[3] = c[3];
+                    Ok(())
+                }
+            }
+        }
+    };
+    (
+    // lum colors (1 component)
+    // - $C      : the type of the inner component
+    // - $BITS   : the number of bits of each inner component
+    // - $INT    : a boolean indicating whether the components are integers
+    // - $LINEAR : a boolean indicating whether it's linear
+    // - $LIGHT  : a boolean indicating whether it's lightness
+    lum: $C:ty, $BITS:literal, $INT:expr, $LINEAR:literal, $LIGHT:literal) => {
+        #[rustfmt::skip]
+        impl $crate::Color for Lum<$C, $LINEAR, $LIGHT> {
+            type Component = $C;
+            const COLOR_BITS: usize = $BITS;
+            const COLOR_COUNT: usize = 1;
+            const COLOR_HAS_ALPHA: bool = false;
+            const COLOR_IS_LINEAR: bool = $LINEAR;
+            const COLOR_IS_INT: bool = $INT;
+            const COLOR_IS_PREMUL: bool = false;
+
+            /// Since the color has no red, the luminosity is returned.
+            fn color_red(&self) -> Self::Component { self.c[0] }
+            /// Since the color has no red, the luminosity is returned.
+            fn color_green(&self) -> Self::Component { self.c[0] }
+            /// Since the color has no red, the luminosity is returned.
+            fn color_blue(&self) -> Self::Component { self.c[0] }
+            /// Since the color has no alpha, the maximum normalized value is returned.
+            fn color_alpha(&self) -> Self::Component { self.c[0] }
+
+            fn color_components_write(&self, b: &mut [$C]) -> Result<(), $crate::NotEnoughSpace> {
+                let needed = Self::COLOR_COUNT;
+                if b.len() < needed {
+                    Err($crate::NotEnoughSpace(Some(Self::COLOR_COUNT)))
+                } else {
+                    b[0] = self.l();
+                    Ok(())
+                }
+            }
+        }
+    };
+}
+
+// impl Color trait
+// type, name, primitive, bits, integer, linear, premul
+impl_color![rgb: Rgb8, u8, 8, true, false];
+impl_color![rgba: Rgba8, u8, 8, true, false, false];
+impl_color![rgba: RgbaPre8, u8, 8, true, false, true];
+impl_color![rgb: Rgb16, u16, 16, true, false];
+impl_color![rgba: Rgba16, u16, 16, true, false, false];
+impl_color![rgba: RgbaPre16, u16, 16, true, false, true];
+crate::items! {
+    impl_color![rgb: RgbF32, f32, 32, false, false];
+    impl_color![rgba: RgbaF32, f32, 32, false, false, false];
+    impl_color![rgba: RgbaPreF32, f32, 32, false, false, true];
+    impl_color![rgb: RgbLinF32, f32, 32, false, true];
+    impl_color![rgba: RgbaLinF32, f32, 32, false, true, false];
+    impl_color![rgba: RgbaLinPreF32, f32, 32, false, true, true];
+}
+crate::items! {
+    impl_color![rgb: RgbF64, f64, 64, false, false];
+    impl_color![rgba: RgbaF64, f64, 64, false, false, false];
+    impl_color![rgba: RgbaPreF64, f64, 64, false, false, true];
+    impl_color![rgb: RgbLinF64, f64, 64, false, true];
+    impl_color![rgba: RgbaLinF64, f64, 64, false, true, false];
+    impl_color![rgba: RgbaLinPreF64, f64, 64, false, true, true];
 }