+++ docs/changelog.md
@@ -183,7 +183,7 @@ Many feature gates are removed in order to make most features make always availa
 - move to [base]:
   - public macros: `CONST!`, `assert_eq_all!`, `assert_approx_eq_all!`, `capture_first!`, `capture_last!`, `capture_tail_tuple!`, `cfg_if!`, `const_assert!`, `define_error!`, `deprecate!`, `enumset!`, `ident_const_index!`, `impl_trait!`, `include_from!`, `is!`, `items!`, `maybe!`, `methods_as_fns!`, `mod_from!`, `sf!`, , `structural_mods!`, `type_marker!`.
   - internal macros: `__dbg!`, `__std!`, `_EMOJI_*`, `_TAG_*`, `_doc!`, `_doc_availability!`, `_doc_miri_warn!`,  `_reexport!`, `_use!`.
-- add tags: `_DOC_*`, `_TAG_[CODEC|CODEGEN_BUILD|CONCURRENCY|DATA|DEBUG|EVENT|EXAMPLE|GEOM_DIR|HASH|ID|PROC_MACRO]`.
+- add tags: `_DOC_*`, `_TAG_[CODEC|CODEGEN_BUILD|COLOR|CONCURRENCY|DATA|DEBUG|EVENT|EXAMPLE|GEOM_DIR|HASH|ID|PROC_MACRO]`.
 - change the emoji for `_TAG_DATA_STRUCTURE`.
 - rename `reexport!` internal macro to `_reexport!`.
   - allow accepting multiple tags.
@@ -251,8 +251,13 @@ Many feature gates are removed in order to make most features make always availa
 ### audio
 - new types: `AudioChannel`, `AudioChannels`.
 
-### font
+### color
+- move to [base]:
+  - types: `Gamma`, `Lum`, `Rgb`, `Rgba`.
+  - aliases: `Lightness`, `LinearLightness`, `Luma`, `Luminance`, `Rgb8`, `Rgba8`, `RgbaPre8`, `Rgb16`, `Rgba16`, `RgbaPre16`, `RgbF32`, `RgbaF32`, `RgbaPreF32`, `RgbF64`, `RgbaF64`, `RgbaPreF64`, `RgbLinF32`, `RgbaLinF32`, `RgbaLinPreF32`, `RgbLinF64`, `RgbaLinF64`, `RgbaLinPreF64`.
+- new type `GammaConst`.
 
+### font
 - new type: `FontArt`.
 - new const: `FONT_ART_3_4`.
 - rename: